Added agenda_check for organisation
