+++ .gitignore
@@ -53,4 +53,6 @@ dkms.conf
 cub3d
 .DS_Store
 libs/.DS_Store
-
+*/.DS_Store
+.vscode/*
+.vscode/launch.json
