+++ makefile
@@ -9,7 +9,8 @@ SRCS	=	./srcs/main.c \
 			./srcs/free.c \
 			./srcs/init.c \
 			./srcs/graphics/colour.c \
-			./srcs/parsing/error_checking.c \
+			./srcs/parsing/agenda_check.c \
+			./srcs/parsing/error_check.c \
 			./srcs/parsing/read_map.c \
 			./srcs/parsing/validate_map.c \
 			./srcs/parsing/is_char_checks.c
